feat: add ts and eslint
